app_streamlit/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
folium>=0.14.0
streamlit-folium>=0.15.0
scikit-learn>=1.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
bs4
sentence-transformers>=2.2.0
selenium
mistralai


